SoftGPU: Move sampler code to a dedicated file.
